+++ packages/client/src/hooks/index.ts
@@ -1,4 +1,10 @@
 export * from './useClientContext';
+export * from './useDebugState';
+export * from './useFeatureFlag';
 export * from './useLibraryContext';
+export * from './useOnboardingStore';
 export * from './useOnlineLocations';
+export * from './useP2PEvents';
 export * from './usePlausible';
+export * from './useTelemetryState';
+export * from './useThemeStore';
