+++ packages/client/src/hooks/useOnlineLocations.ts
@@ -4,6 +4,7 @@ import { useBridgeSubscription } from '../rspc';
 export function useOnlineLocations() {
 	const [state, setState] = useState<number[][] | null>(null);
 
+	// TODO: Move into a context with P2PEvents
 	useBridgeSubscription(['locations.online'], {
 		onData: (d) => setState(d)
 	});
