+++ packages/client/src/hooks/usePlausible.tsx
@@ -1,6 +1,6 @@
 import Plausible, { PlausibleOptions as PlausibleTrackerOptions } from 'plausible-tracker';
 import { useCallback, useEffect, useRef } from 'react';
-import { PlausiblePlatformType, telemetryStore, useDebugState, useTelemetryState } from '../stores';
+import { PlausiblePlatformType, telemetryStore, useDebugState, useTelemetryState } from '..';
 
 /**
  * This should be in sync with the Core's version.
