+++ packages/client/src/hooks/useThemeStore.ts
@@ -1,12 +1,12 @@
 import { useSnapshot } from 'valtio';
-import { valtioPersist } from './util';
+import { valtioPersist } from '..';
 
 export type Themes = 'vanilla' | 'dark';
 
 const themeStore = valtioPersist('sd-theme', {
 	theme: 'dark' as Themes,
 	syncThemeWithSystem: false,
-	hueValue: 235,
+	hueValue: 235
 });
 
 export function useThemeStore() {
