+++ packages/client/src/index.ts
@@ -15,7 +15,6 @@ if (
 	throw new Error('Please ensure you have patched `globalThis` before importing `@sd/client`!');
 
 export * from './hooks';
-export * from './stores';
 export * from './rspc';
 export * from './core';
 export * from './utils';
