+++ packages/client/src/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './passwordStrength';
+export * from './valito';
