--- packages/client/src/stores/index.ts
@@ -1,7 +1,0 @@
-export * from './debugState';
-export * from './themeStore';
-export * from './util';
-export * from './onboardingStore';
-export * from './telemetryState';
-
-// NOTE: Rule of thumb: read from snapshots, mutate the source.
